+++ commands/html5_info_command.go
@@ -39,13 +39,8 @@ func (c *InfoCommand) Execute(args []string) ExecutionStatus {
 	flagSet := flag.NewFlagSet("html5-info", flag.ContinueOnError)
 	flagSet.Parse(args)
 
-	if flagSet.NArg() > 0 {
-		appHostGUIDs := args[len(args)-flagSet.NArg():]
-		return c.GetServiceInfos(appHostGUIDs)
-	}
-
-	ui.Failed("Incorrect number of arguments passed. See [cf html5-info --help] for more detals")
-	return Failure
+	appHostGUIDs := args[len(args)-flagSet.NArg():]
+	return c.GetServiceInfos(appHostGUIDs)
 }
 
 // GetServiceInfos get html5-apps-repo service app-host plan info
@@ -61,17 +56,72 @@ func (c *InfoCommand) GetServiceInfos(appHostGUIDs []string) ExecutionStatus {
 		return Failure
 	}
 
-	ui.Say("Getting information about service instances with app-host-id %s in org %s / space %s as %s...",
-		terminal.EntityNameColor(strings.Join(appHostGUIDs, ", ")),
-		terminal.EntityNameColor(context.Org),
-		terminal.EntityNameColor(context.Space),
-		terminal.EntityNameColor(context.Username))
+	// If no app-host ID passed, get all
+	if len(appHostGUIDs) == 0 {
+		ui.Say("Getting information about all app-host service instances in org %s / space %s as %s...",
+			terminal.EntityNameColor(context.Org),
+			terminal.EntityNameColor(context.Space),
+			terminal.EntityNameColor(context.Username))
+	} else {
+		ui.Say("Getting information about app-host service instances %s in org %s / space %s as %s...",
+			terminal.EntityNameColor(strings.Join(appHostGUIDs, ", ")),
+			terminal.EntityNameColor(context.Org),
+			terminal.EntityNameColor(context.Space),
+			terminal.EntityNameColor(context.Username))
+	}
 
 	// Get HTML5 context
 	html5Context, err := c.GetHTML5Context(context)
 	if err != nil {
 		ui.Failed(err.Error())
 		return Failure
+	}
+
+	// Find app-host service plan
+	log.Tracef("Looking for app-host service plan\n")
+	var appHostServicePlan *models.CFServicePlan
+	for _, plan := range html5Context.HTML5AppsRepoServicePlans {
+		if plan.Name == "app-host" {
+			appHostServicePlan = &plan
+			break
+		}
+	}
+	if appHostServicePlan == nil {
+		ui.Failed("Could not find app-host service plan")
+		return Failure
+	}
+
+	// Get list of service instances of app-host plan
+	log.Tracef("Getting service instances of %s service app-host plan (%+v)\n", html5Context.ServiceName, appHostServicePlan)
+	var appHostServiceInstances []models.CFServiceInstance
+	appHostServiceInstances, err = clients.GetServiceInstances(c.CliConnection, context.SpaceID, []models.CFServicePlan{*appHostServicePlan})
+	if err != nil {
+		ui.Failed("Could not get service instances for app-host plan: %+v", err)
+		return Failure
+	}
+
+	// If no app-host ID passed, get all of them.
+	// Otherwise find and normalize GUID/Name
+	nameMap := make(map[string]string)
+	if len(appHostGUIDs) == 0 {
+		// Collect app-host IDs
+		for _, serviceInstance := range appHostServiceInstances {
+			appHostGUIDs = append(appHostGUIDs, serviceInstance.GUID)
+			nameMap[serviceInstance.GUID] = serviceInstance.Name
+		}
+	} else {
+		log.Tracef("appHostGUIDs before normalization %+v\n", appHostGUIDs)
+		for _, serviceInstance := range appHostServiceInstances {
+			log.Tracef("Service id and name '%s'->'%s'\n", serviceInstance.GUID, serviceInstance.Name)
+			idx := indexOfString(appHostGUIDs, serviceInstance.Name)
+			if idx >= 0 {
+				// Name was passed instead of GUID -> change name to GUID
+				log.Tracef("Converting service name to id '%s'->'%s'\n", appHostGUIDs[idx], serviceInstance.GUID)
+				appHostGUIDs = replaceString(appHostGUIDs, idx, serviceInstance.GUID)
+			}
+			nameMap[serviceInstance.GUID] = serviceInstance.Name
+		}
+		log.Tracef("appHostGUIDs after normalization %+v\n", appHostGUIDs)
 	}
 
 	sizeMap := make(map[string]int)
@@ -154,6 +204,7 @@ func (c *InfoCommand) GetServiceInfos(appHostGUIDs []string) ExecutionStatus {
 			return Failure
 		}
 		infoRecords = append(infoRecords, InfoRecord{
+			AppHostName: nameMap[appHostGUID],
 			AppHostGUID: appHostGUID,
 			SizeLimit:   meta.SizeLimit,
 			Used:        sizeMap[appHostGUID],
@@ -172,9 +223,10 @@ func (c *InfoCommand) GetServiceInfos(appHostGUIDs []string) ExecutionStatus {
 	ui.Say("")
 
 	// Display information about HTML5 applications
-	table := ui.Table([]string{"app-host-id", "used", "size limit", "status", "last changed"})
+	table := ui.Table([]string{"name", "app-host-id", "used", "size limit", "status", "last changed"})
 	for _, infoRecord := range infoRecords {
-		table.Add(infoRecord.AppHostGUID,
+		table.Add(infoRecord.AppHostName,
+			infoRecord.AppHostGUID,
 			getReadableSize(infoRecord.Used),
 			getReadableSize(infoRecord.SizeLimit),
 			infoRecord.Status,
@@ -185,8 +237,16 @@ func (c *InfoCommand) GetServiceInfos(appHostGUIDs []string) ExecutionStatus {
 	return Success
 }
 
+func replaceString(collection []string, idx int, element string) []string {
+	newCollection := collection[0:idx]
+	newCollection = append(newCollection, element)
+	newCollection = append(newCollection, collection[idx+1:]...)
+	return newCollection
+}
+
 // InfoRecord service information record
 type InfoRecord struct {
+	AppHostName string
 	AppHostGUID string
 	SizeLimit   int
 	Used        int